XRD / XRR Fringe Analysis

Searches for fringes in the XRD or XRR spectrum and returns a list with the thickness, a peak table, and two graphs displaying the results.

xrd.FringeAnalysis(df)

Arguments

df

data frame with theta (2q) and I (intensity)

Value

list with thickness, table of peaks, and two ggplot graphs

Author

Thomas Gredig

Examples

  filename <- xrd.getSampleFiles('rasx')
  data <- xrd.import(filename)
  analysis <- xrd.FringeAnalysis(data)
  print(analysis$t.nm)
#>   [1] 110.41967  63.09932 176.68396 126.20552  84.13967  98.16646  73.62775
#>   [8] 110.44586 160.65275 135.94016 103.95736 110.45839  60.94539 160.68098
#>  [15] 126.25272 103.97631 176.76496 103.98250 135.98170  70.71354 103.99556
#>  [22] 252.56775 126.28702  93.05727 126.29706 136.01673  98.23808  98.24253
#>  [29] 147.36942 160.77139  65.50269 126.33345 252.67397 136.05887 126.34500
#>  [36]  88.44562 117.93321 221.13189 147.42542 104.06929  70.77140 176.93751
#>  [43] 117.96268  98.30705 294.93175  65.54366 442.44066 252.82751 117.99019
#>  [50] 160.90222  93.15831 295.01298 252.87351 126.44101 221.27960  93.17444
#>  [57] 110.65099 136.19237 118.03889 136.20513 136.21112 136.21717 177.08931
#>  [64]  88.54925 196.78606  80.50778 196.80761 147.61124 177.14049 196.82952
#>  [71] 110.72163 221.45298 196.85324 126.55389 253.11766 253.12427 221.48996
#>  [78] 295.32773  80.54819  55.38260 354.47393  84.40283  63.30822 136.36727
#>  [85] 295.47364 147.74217 221.62220 104.29809  98.51080 110.83242 221.67587
#>  [92] 136.42190 177.35702 147.80436 295.61999 126.69967 136.45367 354.79325
#>  [99] 253.43031 136.46838 221.77123  98.57059  88.72094 161.32189  77.15980
#> [106] 161.34646 118.32770 161.36549 126.79442  88.76304 136.56896 110.96979
#> [113] 126.83154 221.96672 197.31169  93.46966 118.40454  88.81090 161.48651
#> [120] 161.49518 111.03528 111.04404 177.68192  57.32262 177.71843 222.15815
#> [127] 148.11296 136.72837 136.73747 111.10750  84.66145  98.78174  88.91245
#> [134]  84.68771 161.69032 161.69985 177.87987 177.88947 254.13951 177.90587
#> [141]  71.16915 111.21431  77.37482 254.25277 118.65859 118.66858 161.83266
#> [148] 148.35627 127.17192